refactor(PSGO-268): decompose validation module
Splits the 848-line validation.ts into src/tools/validation/{index,types,sanitize,
validate,model}.ts (all <~400 LOC). Public API at @/tools/validation unchanged
(re-exported via index.ts); storage-schema.json import rewired. Pure move — 30
components tests green, tsc/lint clean.
