--- src/tools/validation.ts
@@ -1,848 +1,0 @@
-/**
- * Validator functions for component configuration data generated by agents.
- *
- * Ported from `tools/validation.py`. There is no `ajv` (or any JSON-schema library)
- * in this project's dependencies, so the validator is hand-written here. It is
- * deliberately *tolerant* of the schema inconsistencies the Keboola Developer Portal
- * UI schemas exhibit (boolean `required`, empty `enum`, `properties: []`, the UI-only
- * `button` type, …) and — matching the Python behaviour — when the schema itself is
- * invalid or absent, validation is skipped (logged + continue) so a broken upstream
- * schema never blocks a write.
- */
-
-import { logger } from '@/logger';
-import storageSchema from './storage-schema.json' with { type: 'json' };
-
-export type JsonDict = Record<string, unknown>;
-export type JsonSchema = Record<string, unknown>;
-
-const SNOWFLAKE_TRANSFORMATION_ID = 'keboola.snowflake-transformation';
-const BIGQUERY_TRANSFORMATION_ID = 'keboola.google-bigquery-transformation';
-
-/** Minimal Component shape the validators need. */
-export type ComponentForValidation = {
-  component_id: string;
-  component_type: string;
-  capabilities: { is_row_based: boolean };
-  configuration_schema?: JsonSchema | null;
-  configuration_row_schema?: JsonSchema | null;
-};
-
-export type ValidationContext = {
-  component_id: string;
-  configuration_id?: string | null;
-  configuration_row_id?: string | null;
-  scope?: 'parameters' | 'storage' | string | null;
-};
-
-const contextToString = (ctx: ValidationContext): string => {
-  let s = `component_id=${ctx.component_id}`;
-  if (ctx.configuration_id) s += `, configuration_id=${ctx.configuration_id}`;
-  if (ctx.configuration_row_id) s += `, configuration_row_id=${ctx.configuration_row_id}`;
-  if (ctx.scope) s += `, scope=${ctx.scope}`;
-  return s;
-};
-
-/**
- * Raised when an instance is invalid under a schema. Carries a recoverable, agent-
- * friendly message — the port of Python's RecoverableValidationError.
- */
-export class RecoverableValidationError extends Error {
-  readonly validator: string;
-  readonly validatorValue: unknown;
-  readonly schemaPath: (string | number)[];
-  readonly instancePath: (string | number)[];
-  readonly instance: unknown;
-  readonly baseMessage: string;
-  initialMessage?: string;
-  validationContext?: ValidationContext;
-
-  constructor(opts: {
-    message: string;
-    validator: string;
-    validatorValue: unknown;
-    schemaPath: (string | number)[];
-    instancePath: (string | number)[];
-    instance: unknown;
-    initialMessage?: string;
-    validationContext?: ValidationContext;
-  }) {
-    super(opts.message);
-    this.name = 'RecoverableValidationError';
-    this.baseMessage = opts.message;
-    this.validator = opts.validator;
-    this.validatorValue = opts.validatorValue;
-    this.schemaPath = opts.schemaPath;
-    this.instancePath = opts.instancePath;
-    this.instance = opts.instance;
-    this.initialMessage = opts.initialMessage;
-    this.validationContext = opts.validationContext;
-    this.message = this.format();
-  }
-
-  private format(): string {
-    let s = `${this.baseMessage}\n`;
-    if (this.validator && this.validatorValue !== undefined && this.validatorValue !== null) {
-      const schemaPath = this.schemaPath.map((p) => `[${JSON.stringify(p)}]`).join('');
-      s += `Failed validating ${JSON.stringify(this.validator)} in schema${schemaPath}:\n`;
-      s += `    ${JSON.stringify({ [this.validator]: this.validatorValue }, null, 2)}\n`;
-    }
-    if (this.instancePath.length > 0) {
-      const instancePath = this.instancePath.map((p) => `[${JSON.stringify(p)}]`).join('');
-      s += `On instance${instancePath}:\n`;
-      s += `    ${JSON.stringify(this.instance, null, 4)}\n`;
-    }
-    if (this.initialMessage) s += `${this.initialMessage}\n`;
-    if (this.validationContext) {
-      s += `Validation component context: ${contextToString(this.validationContext)}\n`;
-    }
-    if (
-      this.validator === 'required' &&
-      Array.isArray(this.validatorValue) &&
-      this.validationContext?.scope === 'parameters'
-    ) {
-      const requiredFields = (this.validatorValue as string[]).map((f) => `\`${f}\``).join(', ');
-      s +=
-        `HINT: Ensure ALL of the following required fields are present in \`parameters\`: ${requiredFields}. ` +
-        'Call `get_components` to retrieve the full schema and `get_config_examples` for real-world examples.' +
-        '\n';
-    }
-    return s.replace(/\s+$/, '');
-  }
-}
-
-/** A schema that is structurally invalid (port of jsonschema.SchemaError). */
-class SchemaError extends Error {}
-
-// ---------------------------------------------------------------------------
-// Schema sanitization — port of KeboolaParametersValidator.sanitize_schema.
-// ---------------------------------------------------------------------------
-
-const isObject = (value: unknown): value is JsonDict =>
-  typeof value === 'object' && value !== null && !Array.isArray(value);
-
-/**
- * Normalizes a JSON schema *in place* and returns it. Mirrors the Python
- * `_sanitize_node`: strips empty `enum`, converts boolean-ish `required` flags into
- * the list form (propagating up to the parent), and turns `properties: []` into `{}`.
- * Returns `[schema, isCurrentRequired]`.
- */
-const sanitizeNode = (schema: unknown): [unknown, boolean | null] => {
-  if (!isObject(schema)) {
-    return [schema, false];
-  }
-
-  if ('enum' in schema && Array.isArray(schema.enum) && schema.enum.length === 0) {
-    delete schema.enum;
-  }
-
-  let isCurrentRequired: boolean | null = null;
-  let required = schema.required;
-  if (!Array.isArray(required)) {
-    if (required !== undefined) {
-      isCurrentRequired = String(required).toLowerCase() === 'true';
-    }
-    required = [];
-  }
-  const requiredList = required as string[];
-
-  let properties = schema.properties;
-  if (properties !== undefined && properties !== null) {
-    if (Array.isArray(properties) && properties.length === 0) {
-      properties = {};
-    } else if (!isObject(properties)) {
-      throw new SchemaError(`properties must be a dictionary, got ${typeof properties}`);
-    }
-    const props = properties as JsonDict;
-    for (const propertyName of Object.keys(props)) {
-      const [sanitized, isChildRequired] = sanitizeNode(props[propertyName]);
-      props[propertyName] = sanitized;
-      if (isChildRequired === true && !requiredList.includes(propertyName)) {
-        requiredList.push(propertyName);
-      } else if (isChildRequired === false && requiredList.includes(propertyName)) {
-        requiredList.splice(requiredList.indexOf(propertyName), 1);
-      }
-    }
-    schema.properties = props;
-  }
-
-  if (requiredList.length > 0) {
-    schema.required = [...requiredList];
-  } else {
-    delete schema.required;
-  }
-
-  if ('items' in schema) {
-    const items = schema.items;
-    if (isObject(items)) {
-      schema.items = sanitizeNode(items)[0];
-    } else if (Array.isArray(items)) {
-      schema.items = items.map((item) => (isObject(item) ? sanitizeNode(item)[0] : item));
-    }
-  }
-
-  for (const keyword of ['allOf', 'anyOf', 'oneOf'] as const) {
-    if (keyword in schema && Array.isArray(schema[keyword])) {
-      schema[keyword] = (schema[keyword] as unknown[]).map((s) =>
-        isObject(s) ? sanitizeNode(s)[0] : s,
-      );
-    }
-  }
-
-  for (const keyword of ['not', 'if', 'then', 'else'] as const) {
-    if (keyword in schema && isObject(schema[keyword])) {
-      schema[keyword] = sanitizeNode(schema[keyword])[0];
-    }
-  }
-
-  if ('additionalProperties' in schema && isObject(schema.additionalProperties)) {
-    schema.additionalProperties = sanitizeNode(schema.additionalProperties)[0];
-  }
-
-  if ('patternProperties' in schema && isObject(schema.patternProperties)) {
-    const pp = schema.patternProperties as JsonDict;
-    for (const pattern of Object.keys(pp)) {
-      if (isObject(pp[pattern])) pp[pattern] = sanitizeNode(pp[pattern])[0];
-    }
-  }
-
-  for (const keyword of ['definitions', '$defs'] as const) {
-    if (keyword in schema && isObject(schema[keyword])) {
-      const defs = schema[keyword] as JsonDict;
-      for (const name of Object.keys(defs)) {
-        if (isObject(defs[name])) defs[name] = sanitizeNode(defs[name])[0];
-      }
-    }
-  }
-
-  return [schema, isCurrentRequired];
-};
-
-const sanitizeSchema = (schema: JsonSchema): JsonSchema =>
-  sanitizeNode(structuredClone(schema))[0] as JsonSchema;
-
-// ---------------------------------------------------------------------------
-// Core validator — a tolerant subset of JSON Schema draft-07.
-// ---------------------------------------------------------------------------
-
-type Fail = (
-  validator: string,
-  validatorValue: unknown,
-  message: string,
-  schemaPath: (string | number)[],
-  instancePath: (string | number)[],
-  instance: unknown,
-) => never;
-
-const jsTypeMatches = (type: string, instance: unknown): boolean => {
-  switch (type) {
-    case 'object':
-      return isObject(instance);
-    case 'array':
-      return Array.isArray(instance);
-    case 'string':
-      return typeof instance === 'string';
-    case 'boolean':
-      return typeof instance === 'boolean';
-    case 'null':
-      return instance === null;
-    case 'number':
-      return typeof instance === 'number';
-    case 'integer':
-      return typeof instance === 'number' && Number.isInteger(instance);
-    case 'button':
-      // UI-only construct accepted as valid (port of check_button_type).
-      return isObject(instance) && instance.type === 'button';
-    default:
-      // Unknown type keyword: be tolerant and accept.
-      return true;
-  }
-};
-
-const deepEqual = (a: unknown, b: unknown): boolean => {
-  if (a === b) return true;
-  if (typeof a !== typeof b) return false;
-  if (Array.isArray(a) && Array.isArray(b)) {
-    return a.length === b.length && a.every((x, i) => deepEqual(x, b[i]));
-  }
-  if (isObject(a) && isObject(b)) {
-    const ka = Object.keys(a);
-    const kb = Object.keys(b);
-    return ka.length === kb.length && ka.every((k) => deepEqual(a[k], b[k]));
-  }
-  return false;
-};
-
-/** Returns true if `instance` is valid under `schema` (no error thrown). */
-const isValid = (schema: unknown, instance: unknown): boolean => {
-  try {
-    validateNode(schema, instance, [], [], () => {
-      throw new Error('invalid');
-    });
-    return true;
-  } catch {
-    return false;
-  }
-};
-
-const validateNode = (
-  schema: unknown,
-  instance: unknown,
-  schemaPath: (string | number)[],
-  instancePath: (string | number)[],
-  fail: Fail,
-): void => {
-  if (typeof schema === 'boolean') {
-    if (!schema)
-      fail('schema', false, 'False schema rejects all values', schemaPath, instancePath, instance);
-    return;
-  }
-  if (!isObject(schema)) return;
-
-  // type
-  if ('type' in schema) {
-    const types = Array.isArray(schema.type) ? (schema.type as string[]) : [schema.type as string];
-    if (!types.some((t) => jsTypeMatches(t, instance))) {
-      fail(
-        'type',
-        schema.type,
-        `${JSON.stringify(instance)} is not of type ${types.map((t) => JSON.stringify(t)).join(', ')}`,
-        [...schemaPath, 'type'],
-        instancePath,
-        instance,
-      );
-    }
-  }
-
-  // enum
-  if ('enum' in schema && Array.isArray(schema.enum)) {
-    if (!schema.enum.some((v) => deepEqual(v, instance))) {
-      fail(
-        'enum',
-        schema.enum,
-        `${JSON.stringify(instance)} is not one of ${JSON.stringify(schema.enum)}`,
-        [...schemaPath, 'enum'],
-        instancePath,
-        instance,
-      );
-    }
-  }
-
-  // const
-  if ('const' in schema && !deepEqual(schema.const, instance)) {
-    fail(
-      'const',
-      schema.const,
-      `${JSON.stringify(instance)} was expected`,
-      [...schemaPath, 'const'],
-      instancePath,
-      instance,
-    );
-  }
-
-  // numeric constraints
-  if (typeof instance === 'number') {
-    if (typeof schema.minimum === 'number' && instance < schema.minimum) {
-      fail(
-        'minimum',
-        schema.minimum,
-        `${instance} is less than the minimum of ${schema.minimum}`,
-        [...schemaPath, 'minimum'],
-        instancePath,
-        instance,
-      );
-    }
-    if (typeof schema.maximum === 'number' && instance > schema.maximum) {
-      fail(
-        'maximum',
-        schema.maximum,
-        `${instance} is greater than the maximum of ${schema.maximum}`,
-        [...schemaPath, 'maximum'],
-        instancePath,
-        instance,
-      );
-    }
-  }
-
-  // string constraints
-  if (typeof instance === 'string') {
-    if (typeof schema.minLength === 'number' && instance.length < schema.minLength) {
-      fail(
-        'minLength',
-        schema.minLength,
-        `${JSON.stringify(instance)} is too short`,
-        [...schemaPath, 'minLength'],
-        instancePath,
-        instance,
-      );
-    }
-    if (typeof schema.maxLength === 'number' && instance.length > schema.maxLength) {
-      fail(
-        'maxLength',
-        schema.maxLength,
-        `${JSON.stringify(instance)} is too long`,
-        [...schemaPath, 'maxLength'],
-        instancePath,
-        instance,
-      );
-    }
-  }
-
-  // array constraints
-  if (Array.isArray(instance)) {
-    if (typeof schema.minItems === 'number' && instance.length < schema.minItems) {
-      fail(
-        'minItems',
-        schema.minItems,
-        `${JSON.stringify(instance)} is too short`,
-        [...schemaPath, 'minItems'],
-        instancePath,
-        instance,
-      );
-    }
-    if (typeof schema.maxItems === 'number' && instance.length > schema.maxItems) {
-      fail(
-        'maxItems',
-        schema.maxItems,
-        `${JSON.stringify(instance)} is too long`,
-        [...schemaPath, 'maxItems'],
-        instancePath,
-        instance,
-      );
-    }
-    const items = schema.items;
-    if (isObject(items) || typeof items === 'boolean') {
-      instance.forEach((item, i) =>
-        validateNode(items, item, [...schemaPath, 'items'], [...instancePath, i], fail),
-      );
-    } else if (Array.isArray(items)) {
-      instance.forEach((item, i) => {
-        if (i < items.length) {
-          validateNode(items[i], item, [...schemaPath, 'items', i], [...instancePath, i], fail);
-        }
-      });
-    }
-  }
-
-  // object constraints
-  if (isObject(instance)) {
-    if (Array.isArray(schema.required)) {
-      const missing = (schema.required as string[]).filter((key) => !(key in instance));
-      if (missing.length > 0) {
-        fail(
-          'required',
-          schema.required,
-          `${JSON.stringify(missing[0])} is a required property`,
-          [...schemaPath, 'required'],
-          instancePath,
-          instance,
-        );
-      }
-    }
-
-    const properties = isObject(schema.properties) ? schema.properties : {};
-    for (const key of Object.keys(properties)) {
-      if (key in instance) {
-        validateNode(
-          properties[key],
-          instance[key],
-          [...schemaPath, 'properties', key],
-          [...instancePath, key],
-          fail,
-        );
-      }
-    }
-
-    const patternProperties = isObject(schema.patternProperties) ? schema.patternProperties : {};
-    const patternKeys = Object.keys(patternProperties);
-    for (const key of Object.keys(instance)) {
-      for (const pattern of patternKeys) {
-        let re: RegExp | null = null;
-        try {
-          re = new RegExp(pattern);
-        } catch {
-          re = null;
-        }
-        if (re && re.test(key)) {
-          validateNode(
-            patternProperties[pattern],
-            instance[key],
-            [...schemaPath, 'patternProperties', pattern],
-            [...instancePath, key],
-            fail,
-          );
-        }
-      }
-    }
-
-    if (isObject(schema.additionalProperties)) {
-      const declared = new Set(Object.keys(properties));
-      for (const key of Object.keys(instance)) {
-        if (declared.has(key)) continue;
-        if (patternKeys.some((p) => safeTest(p, key))) continue;
-        validateNode(
-          schema.additionalProperties,
-          instance[key],
-          [...schemaPath, 'additionalProperties'],
-          [...instancePath, key],
-          fail,
-        );
-      }
-    } else if (schema.additionalProperties === false) {
-      const declared = new Set(Object.keys(properties));
-      for (const key of Object.keys(instance)) {
-        if (declared.has(key)) continue;
-        if (patternKeys.some((p) => safeTest(p, key))) continue;
-        fail(
-          'additionalProperties',
-          false,
-          `Additional properties are not allowed (${JSON.stringify(key)} was unexpected)`,
-          [...schemaPath, 'additionalProperties'],
-          instancePath,
-          instance,
-        );
-      }
-    }
-  }
-
-  // allOf
-  if (Array.isArray(schema.allOf)) {
-    schema.allOf.forEach((sub, i) =>
-      validateNode(sub, instance, [...schemaPath, 'allOf', i], instancePath, fail),
-    );
-  }
-
-  // anyOf
-  if (Array.isArray(schema.anyOf)) {
-    if (!schema.anyOf.some((sub) => isValid(sub, instance))) {
-      fail(
-        'anyOf',
-        schema.anyOf,
-        `${JSON.stringify(instance)} is not valid under any of the given schemas`,
-        [...schemaPath, 'anyOf'],
-        instancePath,
-        instance,
-      );
-    }
-  }
-
-  // oneOf
-  if (Array.isArray(schema.oneOf)) {
-    const matches = schema.oneOf.filter((sub) => isValid(sub, instance)).length;
-    if (matches !== 1) {
-      fail(
-        'oneOf',
-        schema.oneOf,
-        `${JSON.stringify(instance)} is valid under ${matches} of the given schemas (expected exactly 1)`,
-        [...schemaPath, 'oneOf'],
-        instancePath,
-        instance,
-      );
-    }
-  }
-
-  // not
-  if (schema.not !== undefined && isValid(schema.not, instance)) {
-    fail(
-      'not',
-      schema.not,
-      `${JSON.stringify(instance)} is not allowed`,
-      [...schemaPath, 'not'],
-      instancePath,
-      instance,
-    );
-  }
-
-  // if / then / else
-  if (schema.if !== undefined) {
-    if (isValid(schema.if, instance)) {
-      if (schema.then !== undefined) {
-        validateNode(schema.then, instance, [...schemaPath, 'then'], instancePath, fail);
-      }
-    } else if (schema.else !== undefined) {
-      validateNode(schema.else, instance, [...schemaPath, 'else'], instancePath, fail);
-    }
-  }
-};
-
-const safeTest = (pattern: string, value: string): boolean => {
-  try {
-    return new RegExp(pattern).test(value);
-  } catch {
-    return false;
-  }
-};
-
-/**
- * Validates `jsonData` against `schema`. On a validation failure, throws a
- * RecoverableValidationError. On a structurally invalid schema, logs and returns
- * (continue as if valid) — the parity behaviour of `_validate_json_against_schema`.
- */
-const validateJsonAgainstSchema = (opts: {
-  jsonData: unknown;
-  schema: JsonSchema;
-  initialMessage?: string;
-  validationContext?: ValidationContext;
-  sanitize?: boolean;
-}): void => {
-  let schema = opts.schema;
-  try {
-    if (opts.sanitize) schema = sanitizeSchema(opts.schema);
-  } catch (error) {
-    logger.warn({ err: error }, 'The validation schema is not valid; skipping validation.');
-    return;
-  }
-
-  const fail: Fail = (validator, validatorValue, message, schemaPath, instancePath, instance) => {
-    throw new RecoverableValidationError({
-      message,
-      validator,
-      validatorValue,
-      schemaPath,
-      instancePath,
-      instance,
-      initialMessage: opts.initialMessage,
-      validationContext: opts.validationContext,
-    });
-  };
-
-  try {
-    validateNode(schema, opts.jsonData, [], [], fail);
-  } catch (error) {
-    if (error instanceof RecoverableValidationError) throw error;
-    // Treat any non-validation error as an invalid schema → skip (continue).
-    logger.warn({ err: error }, 'The validation schema is not valid; skipping validation.');
-  }
-};
-
-// ---------------------------------------------------------------------------
-// Public validators (ported from validation.py public functions).
-// ---------------------------------------------------------------------------
-
-const STORAGE_VALIDATION_INITIAL_MESSAGE =
-  'The provided storage configuration input does not follow the storage schema.\n';
-const ROOT_PARAMETERS_INITIAL = (componentId: string) =>
-  `The provided Root parameters configuration input does not follow the Root parameter json schema for component ` +
-  `id: ${componentId}.\n`;
-const ROW_PARAMETERS_INITIAL = (componentId: string) =>
-  `The provided Row parameters configuration input does not follow the Row parameter json schema for component ` +
-  `id: ${componentId}.\n`;
-
-const validateStorageConfigurationAgainstSchema = (
-  storage: JsonDict,
-  initialMessage?: string,
-  validationContext?: ValidationContext,
-): JsonDict => {
-  validateJsonAgainstSchema({
-    jsonData: storage,
-    schema: storageSchema as JsonSchema,
-    initialMessage,
-    validationContext,
-  });
-  return storage;
-};
-
-const validateStorageConfiguration = (
-  storage: JsonDict | null | undefined,
-  component: ComponentForValidation,
-  initialMessage: string | undefined,
-  opts: {
-    isRowStorage: boolean;
-    configurationId?: string | null;
-    configurationRowId?: string | null;
-  },
-): JsonDict => {
-  // Normalize to {'storage': storage | {} } — the agent may pass {storage: …} or just the inner object.
-  let storageCfg: JsonDict | null;
-  if (storage) {
-    const inner = (storage as JsonDict).storage;
-    storageCfg = (inner !== undefined ? inner : storage) as JsonDict | null;
-  } else {
-    storageCfg = {};
-  }
-
-  if (storageCfg === null || storageCfg === undefined) {
-    logger.warn(
-      `No "storage" configuration provided for component ${component.component_id} of type ${component.component_type}.`,
-    );
-    storageCfg = {};
-  }
-
-  if (
-    component.component_id === SNOWFLAKE_TRANSFORMATION_ID ||
-    component.component_id === BIGQUERY_TRANSFORMATION_ID
-  ) {
-    if (!storageCfg.input && !storageCfg.output) {
-      throw new Error(
-        `The "storage" must contain either "input" or "output" mappings in the configuration of the SQL ` +
-          `transformation "${component.component_id}".`,
-      );
-    }
-  }
-
-  if (component.component_type === 'writer' && component.capabilities.is_row_based) {
-    if (!opts.isRowStorage && Object.keys(storageCfg).length > 0) {
-      throw new Error(
-        `The "storage" must be empty for root configuration of the writer component ` +
-          `"${component.component_id}" since it is row-based. In this case, storage should only be defined ` +
-          'in its outgoing row configurations.',
-      );
-    } else if (opts.isRowStorage && !storageCfg.input) {
-      throw new Error(
-        `The "storage" must contain "input" mappings for the row configuration of the writer component ` +
-          `"${component.component_id}".`,
-      );
-    }
-  }
-
-  if (component.component_type === 'writer' && !component.capabilities.is_row_based) {
-    if (opts.isRowStorage) {
-      logger.warn(
-        `Validating "storage" for row configuration of non-row-based writer ${component.component_id} is not ` +
-          'semantically correct. Possible cause: agent error or wrong component flag. Proceeding with validation.',
-      );
-    }
-    if (!storageCfg.input) {
-      throw new Error(
-        `The "storage" must contain "input" mappings for the root configuration of the writer component ` +
-          `"${component.component_id}".`,
-      );
-    }
-  }
-
-  const fullInitial = (initialMessage ?? '') + '\n' + STORAGE_VALIDATION_INITIAL_MESSAGE;
-  const validationContext: ValidationContext = {
-    component_id: component.component_id,
-    configuration_id: opts.configurationId,
-    configuration_row_id: opts.configurationRowId,
-    scope: 'storage',
-  };
-  const normalized = validateStorageConfigurationAgainstSchema(
-    { storage: storageCfg },
-    fullInitial,
-    validationContext,
-  );
-  return (normalized.storage ?? {}) as JsonDict;
-};
-
-export const validateRootStorageConfiguration = (
-  storage: JsonDict | null | undefined,
-  component: ComponentForValidation,
-  initialMessage?: string,
-  configurationId?: string | null,
-): JsonDict =>
-  validateStorageConfiguration(storage, component, initialMessage, {
-    isRowStorage: false,
-    configurationId,
-  });
-
-export const validateRowStorageConfiguration = (
-  storage: JsonDict | null | undefined,
-  component: ComponentForValidation,
-  initialMessage?: string,
-  configurationId?: string | null,
-  configurationRowId?: string | null,
-): JsonDict =>
-  validateStorageConfiguration(storage, component, initialMessage, {
-    isRowStorage: true,
-    configurationId,
-    configurationRowId,
-  });
-
-const validateParametersConfiguration = (
-  parameters: JsonDict,
-  schema: JsonSchema | null | undefined,
-  componentId: string,
-  initialMessage: string | undefined,
-  configurationId?: string | null,
-  configurationRowId?: string | null,
-): JsonDict => {
-  // Agent may pass {parameters: …} or just the inner object.
-  const inner = (parameters as JsonDict).parameters;
-  const expected = (inner !== undefined ? inner : parameters) as JsonDict;
-
-  if (!schema || Object.keys(schema).length === 0) {
-    logger.warn(`No schema provided for component ${componentId}, skipping validation.`);
-    return expected;
-  }
-
-  validateJsonAgainstSchema({
-    jsonData: expected,
-    schema,
-    initialMessage,
-    validationContext: {
-      component_id: componentId,
-      configuration_id: configurationId,
-      configuration_row_id: configurationRowId,
-      scope: 'parameters',
-    },
-    sanitize: true,
-  });
-  return expected;
-};
-
-export const validateRootParametersConfiguration = (
-  parameters: JsonDict,
-  component: ComponentForValidation,
-  initialMessage?: string,
-  configurationId?: string | null,
-): JsonDict =>
-  validateParametersConfiguration(
-    parameters,
-    component.configuration_schema,
-    component.component_id,
-    (initialMessage ?? '') + '\n' + ROOT_PARAMETERS_INITIAL(component.component_id),
-    configurationId,
-  );
-
-export const validateRowParametersConfiguration = (
-  parameters: JsonDict,
-  component: ComponentForValidation,
-  initialMessage?: string,
-  configurationId?: string | null,
-  configurationRowId?: string | null,
-): JsonDict =>
-  validateParametersConfiguration(
-    parameters,
-    component.configuration_row_schema,
-    component.component_id,
-    (initialMessage ?? '') + '\n' + ROW_PARAMETERS_INITIAL(component.component_id),
-    configurationId,
-    configurationRowId,
-  );
-
-/**
- * Validates a list of processors against their component schemas. Skips processors
- * with no schema or whose schema is from the template (the `print_hello` marker).
- * Port of `validate_processors_configuration`.
- */
-export const validateProcessorsConfiguration = async (
-  fetchComponent: (componentId: string) => Promise<ComponentForValidation>,
-  processors: JsonDict[],
-  initialMessage?: string,
-): Promise<JsonDict[]> => {
-  for (const processor of processors) {
-    const definition = (processor.definition as JsonDict) ?? {};
-    const processorId = definition.component as string;
-    const processorInfo = await fetchComponent(processorId);
-
-    const schema = processorInfo.configuration_schema;
-    if (!schema) continue;
-    const required = Array.isArray((schema as JsonDict).required)
-      ? ((schema as JsonDict).required as string[])
-      : [];
-    if (required.includes('print_hello')) continue;
-
-    validateJsonAgainstSchema({
-      jsonData: processor.parameters,
-      schema,
-      initialMessage: `${initialMessage}\nThe configuration of "${processorId}" processor is not valid.`,
-      validationContext: { component_id: processorId },
-      sanitize: true,
-    });
-  }
-  return processors;
-};
-
-// Exported for unit testing.
-export const __testing = { sanitizeSchema, validateJsonAgainstSchema };
